Break implicit stencils tests into utils and run
